+++ R/showDigitizedImage.R
@@ -70,61 +70,108 @@ showDigitizedImage <- function(nms,deviceType,
   ##   filename otherwise process the filename(s)
   if (inherits(nms,"RFishBC")) nms <- nms$datanm                         # nocov
     else nms <- iHndlFilenames(nms,filter="RData",multi=TRUE)
-  ## Prepare for multiple readings #############################################
-  # check that all were created with digitizeRadii() and from same image
-  tmp <- NULL
-  for (i in seq_along(nms)) {
-    if (!isRData(nms[i]))
-      STOP("File is not an RData file saved from 'digitizeRadii().")
-    dat <- readRDS(nms[i])
-    if (!inherits(dat,"RFishBC"))
-      STOP("File does not appear to be from 'digitizeRadii().")
-    tmp <- c(tmp,dat$image)
-  }
-  if (length(unique(tmp))>1)
-    STOP("Files appear to derive from different structure images.")
-  # expand pchs, colors, cexs, lwds to number of transects
+  
+  ## Get number of readings ####################################################
   num2do <- length(nms)
-  pch.show <- rep(pch.show,ceiling(num2do/length(pch.show)))
-  col.show <- rep(col.show,ceiling(num2do/length(col.show)))
-  cex.show <- rep(cex.show,ceiling(num2do/length(cex.show)))
-  col.connect <- rep(col.connect,ceiling(num2do/length(col.connect)))
-  lwd.connect <- rep(lwd.connect,ceiling(num2do/length(lwd.connect)))
 
-  ## Display results ###########################################################
-  for (i in seq_along(nms)) {
-    dat <- readRDS(nms[i])
-    #### If first then show the image
-    if (i==1) {
-      iGetImage(dat$image,id=NULL,
-                windowSize=dat$windowSize,deviceType=deviceType,
-                showInfo=FALSE,pos.info=NULL,cex.info=NULL,col.info=NULL)
-      origImage <- dat$image
-      #### Show scale-bar, if it was digitized, and scale-bar length if asked
-      #### for (but only for the first reading)
-      if (!is.null(dat$sbPts)) {
-        graphics::lines(y~x,data=dat$sbPts,col=col.scaleBar,lwd=lwd.scaleBar)
-        if (showScaleBarLength) iShowScaleBarLength(dat,col.scaleBar,cex.scaleBar)
+  ## Show the reading(s) ... handle one and multiple readings differently ######
+  if (num2do==1) {
+    ### Make sure file is an RData file from digitizeRadii()
+    if (!isRData(nms)) 
+      STOP("File is not an RData file saved from 'digitizeRadii().")
+    dat <- readRDS(nms)
+    if (!inherits(dat,"RFishBC")) 
+      STOP("File does not appear to be from 'digitizeRadii().")
+    iShowOneDigitizedImage(dat,deviceType,
+                           showScaleBarLength,col.scaleBar,lwd.scaleBar,cex.scaleBar,
+                           connect,col.connect,lwd.connect,
+                           pch.show,col.show,cex.show,
+                           showAnnuliLabels,annuliLabels,col.ann,cex.ann)
+  } else {
+    tmp <- NULL
+    for (i in seq_along(nms)) {
+      ### Make sure each file is an RData file from digitizeRadii()
+      if (!isRData(nms[i]))
+        STOP(nms[i],"is not an RData file saved from 'digitizeRadii().")
+      dat <- readRDS(nms[i])
+      if (!inherits(dat,"RFishBC"))
+        STOP(nms[i],"does not appear to be from 'digitizeRadii().")
+      if (!is.null(tmp)) {
+        if (dat$image!=tmp) {
+          grDevices::dev.off()
+          STOP("Files appear to derive from different structure images.")
+        }
       }
-    }
-    #### Show connected points if asked to do so
-    if (connect) graphics::lines(y~x,data=dat$pts,
-                                 lwd=lwd.connect[i],col=col.connect[i])
-    #### Show points (only show points that were considered annuli)
-    graphics::points(dat$pts[2:(nrow(dat$pts)-1),],
-                     pch=pch.show[i],col=col.show[i],cex=cex.show[i])
-    if (dat$edgeIsAnnulus) ## add on edge if considered an annulus
-      graphics::points(dat$pts[nrow(dat$pts),],
-                       pch=pch.show[i],col=col.show[i],cex=cex.show[i])
-    #### Show annuli labels if asked to do so AND only if one set of readings
-    if (num2do==1) {
-      #### Show annuli labels if asked to do so
-      if (showAnnuliLabels) iShowAnnuliLabels(dat,annuliLabels=annuliLabels,
-                                              col.ann=col.ann,cex.ann=cex.ann)
+      tmp <- dat$image
+      
+      ### expand pchs, colors, cexs, lwds to number of readings/transects
+      pch.show <- rep(pch.show,ceiling(num2do/length(pch.show)))
+      col.show <- rep(col.show,ceiling(num2do/length(col.show)))
+      cex.show <- rep(cex.show,ceiling(num2do/length(cex.show)))
+      col.connect <- rep(col.connect,ceiling(num2do/length(col.connect)))
+      lwd.connect <- rep(lwd.connect,ceiling(num2do/length(lwd.connect)))
+      
+      ### Make or add to image
+      if (i==1) iShowOneDigitizedImage(dat,deviceType,
+                                       showScaleBarLength,col.scaleBar,
+                                       lwd.scaleBar,cex.scaleBar,
+                                       connect,col.connect[i],lwd.connect[i],
+                                       pch.show[i],col.show[i],cex.show[i],
+                                       showAnnuliLabels=FALSE,annuliLabels="",
+                                       col.ann[i],cex.ann[i])
+      else {
+        ## Show connected points if asked to do so
+        if (connect) graphics::lines(y~x,data=dat$pts,
+                                     lwd=lwd.connect[i],col=col.connect[i])
+        ## Show points (only show points that were considered annuli)
+        graphics::points(dat$pts[2:(nrow(dat$pts)-1),],
+                         pch=pch.show[i],col=col.show[i],cex=cex.show[i])
+        #### Add on edge if considered an annulus
+        if (dat$edgeIsAnnulus)
+          graphics::points(dat$pts[nrow(dat$pts),],
+                           pch=pch.show[i],col=col.show[i],cex=cex.show[i])
+      }
     }
   }
 }
 
+########################################################################
+## Show the first digitized image.
+########################################################################
+iShowOneDigitizedImage <- function(dat,deviceType,
+                                   showScaleBarLength,col.scaleBar,
+                                   lwd.scaleBar,cex.scaleBar,
+                                   connect,col.connect,lwd.connect,
+                                   pch.show,col.show,cex.show,
+                                   showAnnuliLabels,annuliLabels,col.ann,cex.ann) {
+  ## Get and display the image
+  iGetImage(dat$image,id=NULL,
+            windowSize=dat$windowSize,deviceType=deviceType,
+            showInfo=FALSE,pos.info=NULL,cex.info=NULL,col.info=NULL)
+  ## Show scale-bar, if it was digitized, and scale-bar length if asked for
+  if (!is.null(dat$sbPts)) {
+    graphics::lines(y~x,data=dat$sbPts,col=col.scaleBar,lwd=lwd.scaleBar)
+    if (showScaleBarLength) iShowScaleBarLength(dat,col.scaleBar,cex.scaleBar)
+  }
+  ## Connected points with a line if asked to do so
+  if (connect) graphics::lines(y~x,data=dat$pts,col=col.connect,lwd=lwd.connect)
+  ## Show points (only show points that were considered annuli)
+  num.ann <- ifelse(dat$edgeIsAnnulus,nrow(dat$pts)-1,nrow(dat$pts)-2)
+  if (length(pch.show)>1 & length(pch.show)<num.ann)
+    WARN("'pch.show' was recycled")
+  if (length(col.show)>1 & length(col.show)<num.ann)
+    WARN("'col.show' was recycled")
+  if (length(cex.show)>1 & length(cex.show)<num.ann)
+    WARN("'cex.show' was recycled")
+  graphics::points(dat$pts[2:(nrow(dat$pts)-1),],pch=pch.show,col=col.show,cex=cex.show)
+  #### Add on edge if considered an annulus
+  if (dat$edgeIsAnnulus)
+    graphics::points(dat$pts[nrow(dat$pts),],pch=pch.show[length(pch.show)],
+                     col=col.show[length(col.show)],cex=cex.show[length(cex.show)])
+  #### Show annuli labels if asked to do so
+  if (showAnnuliLabels)
+    iShowAnnuliLabels(dat,annuliLabels=annuliLabels,col.ann=col.ann,cex.ann=cex.ann)
+}
 
 
 
@@ -164,10 +211,10 @@ iShowAnnuliLabels <- function(dat,annuliLabels,col.ann,cex.ann) { # nocov start
   pts <- pts[rownames(pts) %in% annuliLabels,]
   ## Check colors
   if (length(col.ann)>1 & length(col.ann)<length(annuliLabels))
-    WARN("Colors in 'col.ann' will be recylced.")
+    WARN("'col.ann' was recycled.")
   ## Check cexs
   if (length(cex.ann)>1 & length(cex.ann)<length(annuliLabels))
-    WARN("Values in 'cex.ann' will be recylced.")
+    WARN("'cex.ann' was recycled.")
   #### put the labels on the plot
   graphics::text(y~x,data=pts,labels=annuliLabels,font=2,
                  col=col.ann,cex=cex.ann,pos=pos)