+++ NEWS.md
@@ -3,7 +3,7 @@
 * `combineData()`: Modified. Fixed two bugs related to how age-0 fish with no measured annuli are combined (thanks to Rory Feeney). Fixed how returned results are sorted (by id, reading, and ann; rather than just id and ann).
 * `digitizeRadii()`: Modified. Added `sbUnits=` for addressing [#36](https://github.com/droglenc/RFishBC/issues/36). Fixed poor warning message when `snap2transect=TRUE` and `makeTransect=FALSE`.
 * `RFBCoptions()`: Modified. Added `showScaleBarLength=`, `scaleBarUnits=`, and `cex.scaleBar=`.
-* `showDigitizedImage()`: Modified. Added `showScaleBarLength=` and `cex.scaleBar=` with associated code to show the units of the scale-bar beneath the scale-bar (if it was digitized). This addresses [#36](https://github.com/droglenc/RFishBC/issues/36). Changed how annuli labels are printed such that `col.ann=` and `cex.ann` can be appropriate vectors of colors and sizes. Made sure that all values in `annuliLabels=` exist in the data.
+* `showDigitizedImage()`: Modified. Rewrote the function to be less redundant; allow `pch.show=`, `col.show=`, and `cex.show=` to appropriate vectors (when only one reading is displayed); assure that all values in `annuliLabels=` exist in the data; allow `col.ann=` and `cex.ann` to be appropriate vectors; and show units of the scale-bar beneath the scale-bar (if it was digitized) by adding `showScaleBarLength=` and `cex.scaleBar=` (addresses [#36](https://github.com/droglenc/RFishBC/issues/36)).  
 
 # RFishBC 0.2.0 18-Dec-18
 * Added several items (mostly images) to .Rbuildignore to lower bloat of package.